Tensor2Tensor demonstration
This is the first attempt to train a model of polsih language.
This is just a demonstration of tensor2tensor library.
calling gen_data.sh and train.sh should result in a
successfull training.

Data is taken from PolEval 2018 task3 segmented,
small.txt is the first 10000 lines of this file.